+++ src/App.jsx
@@ -11,11 +11,10 @@ import List from "./pages/List";
 import Login from "./pages/Login";
 import Register from "./pages/Register";
 import Schedule from "./pages/Schedule";
-import { AuthContextProvider } from "./context/AuthContext";
 
 function App() {
   return (
-    <AuthContextProvider>
+    <>
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<RootLayout />}>
@@ -33,8 +32,7 @@ function App() {
           </Route>
         </Routes>
       </BrowserRouter>
-    </AuthContextProvider>
+    </>
   );
 }
-
 export default App;
